requirements.txt
Django==5.0.4
# django-allauth==0.61.1
python-dotenv==1.0.1
mysqlclient==2.2.4
django-simple-history==3.5.0
waitress==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pandas==2.2.2
django-oauth-toolkit==2.4.0
requests-mock==1.12.1
crispy-bootstrap4==2024.1
django-crispy-forms==2.1
drf-yasg==1.21.7
setuptools==74.1.2
